Create TCP Sniffer.py
Here it prints out all the incoming and outgoing packets, and with scapy you can't set a direction to just receive the packets, but if I want to only save/print the incoming packets I can do that with a simple if statement (if packet.dst = computer's IP address -> save/print the packet). Now with the macOS I can't get the computer's IP address (for example mine) because of "Root privileges" to do this "if statement". So I'm not sure what route to take regarding this issue. 
I will upload my code for getting the local ip address shortly.